github, google, kakao login 파일 수정
