+++ src/components/LoginForm/GithubLogin.jsx
@@ -2,15 +2,9 @@ import React from 'react';
 import styled from 'styled-components';
 
 const GithubLogin = () => {
-	const uri = `https://github.com/login/oauth/authorize?client_id=c626d47b6711b219e86c&redirect_uri=http://localhost:3000/redirect`;
-	const handleLinkClick = e => {
-		e.preventDefault();
-	};
-	return (
-		<GithubLoginStyled href={uri} onClick={handleLinkClick}>
-			Github
-		</GithubLoginStyled>
-	);
+	const uri = `http://52.79.202.229:8080/oauth2/authorize/github`;
+
+	return <GithubLoginStyled href={uri}>Github</GithubLoginStyled>;
 };
 
 const GithubLoginStyled = styled.a`
